src/requirements.txt
munch==2.5.0
ruamel.yaml==0.16.13
torch==1.3.1
torchvision==0.4.2
numpy==1.17.4
Pillow==6.2.1
h5py==3.2.1
jsonlines==2.0.0
tensorboard==2.0.2
tensorboardX==1.9
tqdm==4.20.0
